feat: add state subscription client to DaemonClient
